Added a new file that contains a script performing yearly projected savings calculation in python

+++ python_introduction/finance_calculator.py
@@ -3,8 +3,8 @@
 income = float(input("Enter your monthly income: "))
 expenses = float(input("Enter your total monthly expenses: "))
 
-savings = income - expenses
-projected_savings = savings * 12 + (savings * 12 * 0.05)
+savings = float(income - expenses)
+projected_savings = float(savings * 12 + (savings * 12 * 0.05))
 
 print("Your monthly savings are: $"+str(savings))
 print("Projected savings after one year, with interest, is: $"+str(projected_savings))
